can only delete menu when there is no orders order this menu

+++ pos_hunsa/app/controllers/menu-delete.php
@@ -9,22 +9,30 @@ $row = $menuClass->first(['menu_id' => $id]);
 
 if ($_SERVER['REQUEST_METHOD'] == "POST" && $row) {
 
-	$menuClass->delete($row['menu_id'], 'menu_id');
+	$orderClass = new Orders;
+	$order_exist = $orderClass->first(['menu_id'=>$id]);
 
-	//delete old image
-	if (file_exists($row['menu_img'])) {
-		unlink($row['menu_img']);
-		
-		//delete cropped image
-		$exp = explode('.', $row['menu_img']);
-		$crop = $exp[0] . '_cropped.' . $exp[1];
-		if (file_exists($crop)){
-			unlink(($crop));
+	//if menu is currently in order then cannot delete
+	if (!$order_exist) {
+
+		$menuClass->delete($row['menu_id'], 'menu_id');
+
+		//delete old image
+		if (file_exists($row['menu_img'])) {
+			unlink($row['menu_img']);
+
+			//delete cropped image
+			$exp = explode('.', $row['menu_img']);
+			$crop = $exp[0] . '_cropped.' . $exp[1];
+			if (file_exists($crop)) {
+				unlink(($crop));
+			}
+
 		}
+		redirect('admin&tab=menu');
+	} else {
+		$errors['error'] = 'This menu is currently in order(s).<br> Please try again later.';
 	}
-
-
-	redirect('admin&tab=menu');
 }
